Repo and service implementations

+++ blogSite.DataAccess/Repositories/Base/EFCommentRepository.cs
@@ -9,14 +9,14 @@ public class EFCommentRepository : ICommentRepository
 {
     private readonly BlogProjectDbContext context;
 
-    public EFCommentRepository(BlogProjectDbContext context)
-        => this.context = context;
+    public EFCommentRepository(BlogProjectDbContext context) =>
+        this.context = context;
 
-    public async Task<IList<Comment>> GetAllAsync() 
-        => await context.Comments.ToListAsync();
+    public async Task<IList<Comment>> GetAllAsync() => 
+        await context.Comments.ToListAsync();
 
-    public async Task<Comment?> GetAsync(int id) 
-        => await context.Comments.FindAsync(id);
+    public async Task<Comment?> GetAsync(int id) => 
+        await context.Comments.FindAsync(id);
 
     public async Task<int> Add(Comment entity)
     {
@@ -34,16 +34,22 @@ public class EFCommentRepository : ICommentRepository
         entity.Updated = DateTime.Now;
 
         context.Comments.Update(entity);
-        return await context.SaveChangesAsync();
+        
+        var affectedRows = await context.SaveChangesAsync();
+        return affectedRows;
     }
 
-    public async Task DeleteAsync(int id)
+    public async Task<int> DeleteAsync(int id)
     {
         var entity = await context.Comments
             .FirstOrDefaultAsync(x => x.Id == id);
 
+        if(entity == null) return 0;
+
         context.Comments.Remove(entity);
-        await context.SaveChangesAsync();
+        
+        var affectedRows = await context.SaveChangesAsync();
+        return affectedRows;
     }
 
     public async Task<bool> IsExist(int id)
