+++ blogSite.DataAccess/Repositories/Base/EFUserRepository.cs
@@ -9,14 +9,14 @@ public class EFUserRepository : IUserRepository
 {
     private readonly BlogProjectDbContext context;
 
-    public EFUserRepository(BlogProjectDbContext context)
-        => this.context = context;
+    public EFUserRepository(BlogProjectDbContext context) =>
+        this.context = context;
 
-    public async Task<IList<User>> GetAllAsync()
-        => await context.Users.ToListAsync();
+    public async Task<IList<User>> GetAllAsync() =>
+        await context.Users.ToListAsync();
 
-    public async Task<User?> GetAsync(int id)
-        => await context.Users.FindAsync(id);
+    public async Task<User?> GetAsync(int id) =>
+        await context.Users.FindAsync(id);
 
     public async Task<int> Add(User user)
     {
@@ -32,13 +32,18 @@ public class EFUserRepository : IUserRepository
         return await context.SaveChangesAsync();
     }
 
-    public async Task DeleteAsync(int id)
+    public async Task<int> DeleteAsync(int id)
     {
+        int affected = 0;
         var user = await context.Users
             .FirstOrDefaultAsync(user => user.Id == id);
+        if(user != null)
+        {
+            context.Users.Remove(user);
+            affected = await context.SaveChangesAsync();
+        }
 
-        context.Users.Remove(user);
-        await context.SaveChangesAsync();
+        return affected;
     }
 
     public async Task<bool> IsExist(int id)
